+++ feature/home/src/main/java/com/selfbell/home/ui/HomeScreen.kt
@@ -42,11 +42,10 @@ fun HomeScreen(
     userProfileName: String,
     criminalMarkers: List<MapMarkerData>,     // 범죄자 마커 리스트
     safetyBellMarkers: List<MapMarkerData>,   // 안심벨 마커 리스트
+
     searchText: String,
     onSearchTextChange: (String) -> Unit,
     onSearchClick: () -> Unit,
-    // alertList: List<AlertItem>, // << 제거
-    // onAlertClick: (AlertItem) -> Unit, // << 제거
     onModalMarkerItemClick: (MapMarkerData) -> Unit, // AddressSearchModal의 아이템 클릭 콜백
     searchedLatLng: LatLng?,
     onMsgReportClick: () -> Unit
@@ -148,42 +147,6 @@ fun HomeScreen(
         )
     }
 }
-
-
-// MapInfoBalloon Composable (별도 파일 또는 HomeScreen 하단에 정의)
-@Composable
-fun MapInfoBalloon(
-    modifier: Modifier = Modifier,
-    address: String,
-    latLng: LatLng, // 마커 위치 정보 (필요시 사용)
-    onDismissRequest: () -> Unit
-) {
-    // Popup을 사용하여 지도 위에 오버레이 형태로 표시
-    // Popup의 위치는 직접 계산하거나, Alignment를 사용할 수 있습니다.
-    // 여기서는 간단하게 화면 중앙 근처에 나타나도록 합니다. (실제로는 마커 위치 기반으로 조정 필요)
-    Popup(
-        alignment = Alignment.Center, // 또는 다른 정렬, offset 사용 가능
-        onDismissRequest = onDismissRequest
-    ) {
-        Surface(
-            modifier = modifier
-                .wrapContentSize()
-                .shadow(4.dp, RoundedCornerShape(8.dp))
-                .clip(RoundedCornerShape(8.dp))
-                .background(Color.White)
-                .clickable(onClick = onDismissRequest), // 말풍선 클릭 시 닫기
-            color = Color.White // Surface 자체 색상
-        ) {
-            Column(modifier = Modifier.padding(16.dp)) {
-                Text(text = "주소", style = Typography.labelSmall)
-                Text(text = address, style = Typography.bodyLarge)
-                Spacer(modifier = Modifier.height(8.dp))
-                // 필요시 추가 정보 (예: "상세보기 버튼 등")
-            }
-        }
-    }
-}
-
 
 fun addOrUpdateMarker(
     naverMap: NaverMap,
